make target_get_serial 64 bit
64 bits of unique information are usually provided by the manufacturer at the factory.

+++ inc/core/codal_target_hal.h
@@ -46,7 +46,7 @@ DEALINGS IN THE SOFTWARE.
 
     int target_random(int max);
 
-    uint32_t target_get_serial();
+    uint64_t target_get_serial();
 
     void target_wait_for_event();
